recored missing cassettes and rename UnknownHttpError
