+++ lib/easypost/errors.rb
@@ -5,7 +5,7 @@ end
 
 require_relative 'errors/easy_post_error'
 require_relative 'errors/end_of_pagination_error'
-require_relative 'errors/external_api_error'
+require_relative 'errors/api/external_api_error'
 require_relative 'errors/filtering_error'
 require_relative 'errors/invalid_object_error'
 require_relative 'errors/invalid_parameter_error'
@@ -28,4 +28,4 @@ require_relative 'errors/api/service_unavailable_error'
 require_relative 'errors/api/ssl_error'
 require_relative 'errors/api/timeout_error'
 require_relative 'errors/api/unauthorized_error'
-require_relative 'errors/api/unknown_http_error'
+require_relative 'errors/api/unknown_api_error'
