+++ lib/easypost/errors/api/external_api_error.rb
