+++ lib/easypost/errors/api/unknown_api_error.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+require_relative 'api_error'
+
+class EasyPost::Errors::UnknownApiError < EasyPost::Errors::ApiError
+end
