--- lib/easypost/errors/api/unknown_http_error.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-require_relative 'api_error'
-
-class EasyPost::Errors::UnknownHttpError < EasyPost::Errors::ApiError
-end
